itinerary creation component now connected to database, properly creates a stored itinerary with title.
