+++ frontend/src/services/itinerary-service.ts
@@ -1,16 +1,22 @@
 import apiClient from './api-client';
 
-export async function postItinerary(entryData: any) {
+interface ItineraryEntry {
+    title: string;
+    description: string;
+  }
+  
+  export async function postItinerary(entryData: ItineraryEntry) {
     const { title, description } = entryData;
-
-    const bodyToSend = {
-        title,
-        description,
-      }      
-
-  const response = await apiClient.post('/itineraries', bodyToSend);
-  return response.data;
-}
+    const bodyToSend = { title, description };
+  
+    try {
+      const response = await apiClient.post('/itineraries', bodyToSend);
+      return response.data;
+    } catch (error) {
+      console.error('Error posting itinerary:', error);
+      throw error;
+    }
+  }  
 
 export async function getItineraries() {
     const response = await apiClient.get('/itineraries');
